Move versions tab/page rendering to templates/package_versions.dart
