+++ app/test/frontend/templates_test.dart
@@ -23,6 +23,7 @@ import 'package:pub_dartlang_org/frontend/templates/layout.dart';
 import 'package:pub_dartlang_org/frontend/templates/listing.dart';
 import 'package:pub_dartlang_org/frontend/templates/misc.dart';
 import 'package:pub_dartlang_org/frontend/templates/package_analysis.dart';
+import 'package:pub_dartlang_org/frontend/templates/package_versions.dart';
 
 import '../shared/html_validation.dart';
 import 'utils.dart';
@@ -440,7 +441,7 @@ void main() {
     });
 
     test('package versions page', () {
-      final String html = templateService.renderPkgVersionsPage(
+      final String html = renderPkgVersionsPage(
         'foobar',
         [
           testPackageVersion,
